+++ packages/core/src/defaults/variants/container-queries.ts
@@ -1,0 +1,7 @@
+export const containerQueries = [
+	{ prefix: "@sm", value: "@container (width >= 40rem)" },
+	{ prefix: "@md", value: "@container (width >= 48rem)" },
+	{ prefix: "@lg", value: "@container (width >= 64rem)" },
+	{ prefix: "@xl", value: "@container (width >= 80rem)" },
+	{ prefix: "@xxl", value: "@container (width >= 96rem)" },
+] as const;